--- a/src/main/php/peer/server/Server.class.php
+++ b/src/main/php/peer/server/Server.class.php
@@ -44,7 +44,7 @@ class Server {
   public function init() {
     $this->socket->create();
     $this->socket->bind(true);
-    $this->socket->listen();
+    $this->socket->listen(SOMAXCONN);
   }
   
   /**